Delete RedisStream::Client class
All code from Client was moved to RedisStream.
