+++ lib/redis_stream.rb
@@ -2,12 +2,20 @@ require 'redis'
 require 'redis_stream/version'
 require 'redis_stream/group'
 require 'redis_stream/stream'
-require 'redis_stream/client'
 
 class RedisStream
   class Error < StandardError; end
 
   def initialize(redis: Redis.new)
-    Client.new(redis)
+    @redis = redis
+  end
+
+  def stream(key)
+    Stream.new(redis: @redis, key: key)
+  end
+
+  def group(key:, name:)
+    new_group = Group.new(redis: @redis, key: key, name: name)
+    new_group.create
   end
 end
